refactor: consolidate frontend backend-facing types and API client
Extract every interface/type alias mirroring a backend JSON shape (across
the admin, search, leagues, and oauth API modules) into a single
src/api/types.ts, and merge admin.ts/leagues.ts/search.ts/oauth.ts into
one src/api/api.ts. One place for the shapes, one place for the fetch
wrappers, instead of four API modules each defining their own types
inline. Manually maintained for now; a natural spot to swap in generated
types later.
