+++ src/FplBot/Services/WebApi/ClientApp/src/views/admin/SlackBroadcastView.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { ref } from "vue";
-import { broadcastToSlack } from "../../api/admin";
+import { broadcastToSlack } from "../../api/api";
 import { describeAdminError } from "../../composables/useAdminAuth";
 
 const message = ref("");
